requirements.txt
psycopg2-binary
python-dotenv
firebase-admin
google-cloud-firestore
boto3
